ppt_translator/tools/rebuilder.py
"""
Rebuilder tool — translates an extraction JSON and writes a new PPTX.

Workflow
--------
1. Read the extraction JSON produced by extract_pptx.
2. Batch-translate all encoded paragraphs via Azure OpenAI.
3. Re-open the original PPTX and traverse shapes in the SAME order as the
   extractor (guaranteed by reusing visit_slide from _common).
4. Apply translated text run-by-run, preserving all formatting.
5. Save the translated presentation.

📝 TASKS (Step 5 — Advanced)
-----------------------------
Implement the two functions below:
  a) `_translate_batch`  — one Azure OpenAI call that translates N paragraphs.
  b) `rebuild_pptx`      — orchestrates everything end-to-end.
"""

from __future__ import annotations

import json
import os
import re
from pathlib import Path
from typing import Any

from openai import AzureOpenAI
from pptx import Presentation

from ._common import decode_paragraph, visit_slide

# Each line sent to the LLM is prefixed with this token so we can parse the
# response back into a list ordered by index.
_PARA_PREFIX = "PARA_"


# ── batch translation ─────────────────────────────────────────────────────────

def _translate_batch(
    client: AzureOpenAI,
    deployment: str,
    encoded_paragraphs: list[str],
    target_language: str,
) -> list[str]:
    """
    Translate a list of encoded paragraph strings in a single API call.

    Protocol
    --------
    • Build a prompt where every paragraph is on its own line prefixed with
      its index:
          PARA_0: [[0]]Hello[[/0]]
          PARA_1: Some plain text
          …

    • Instruct the model (via the system message) to:
        1. Keep the PARA_N: prefix on every output line.
        2. Preserve [[N]]…[[/N]] markers EXACTLY.
        3. Translate only the text inside the markers (or bare text).
        4. Return ONLY the translated paragraphs, nothing else.

    • Parse the response: split on newlines, find lines that start with
      PARA_<int>:, extract the index and the translated content.

    • Return a list in the same order as the input.  If the model omitted
      a paragraph, fall back to the original encoded string for that index.

    Parameters
    ----------
    client              : AzureOpenAI client
    deployment          : Azure OpenAI deployment name
    encoded_paragraphs  : list of encoded strings from extract_pptx
    target_language     : e.g. "French", "German", "Japanese"

    Returns
    -------
    list[str] — translated strings, same length and order as input.

    📝 TODO (Step 5a): Implement this function.
    """
    # TODO: implement _translate_batch
    raise NotImplementedError("_translate_batch is not implemented yet.")


# ── public API ────────────────────────────────────────────────────────────────

def rebuild_pptx(
    input_path: str,
    extraction_json_path: str,
    output_path: str,
    target_language: str,
    batch_size: int = 30,
) -> dict[str, Any]:
    """
    Translate the paragraphs in *extraction_json_path* and write a new PPTX
    to *output_path*, preserving every formatting run from *input_path*.

    Parameters
    ----------
    input_path           : original source .pptx (for layout / formatting)
    extraction_json_path : JSON produced by extract_pptx
    output_path          : where to save the translated .pptx
    target_language      : e.g. "French", "German", "Japanese"
    batch_size           : paragraphs per Azure OpenAI call (default 30)

    Returns
    -------
    On success:
        {
            "status": "success",
            "output_path": <str>,
            "target_language": <str>,
            "slides": <int>,
            "paragraphs_translated": <int>,
            "api_calls": <int>,
        }
    On error:
        {"error": <message>}

    📝 TODO (Step 5b): Implement this function.

    Suggested steps:
      1. Validate input_path and extraction_json_path exist.
      2. Load the extraction JSON; extract the "paragraphs" list and
         "include_notes" flag.
      3. Create an AzureOpenAI client from environment variables:
             AZURE_API_KEY, AZURE_API_BASE, AZURE_API_VERSION,
             AZURE_OPENAI_DEPLOYMENT.
      4. Loop over encoded strings in chunks of `batch_size`, calling
         `_translate_batch` for each chunk.  Collect all results in order.
      5. Re-open the original PPTX.  Use visit_slide + decode_paragraph
         (iterate over translated_list with iter/next) to apply translations.
      6. Save the presentation to output_path (create parents if needed).
      7. Return the summary dict.
    """
    # TODO (Step 5b): Implement rebuild_pptx.
    raise NotImplementedError("rebuild_pptx is not implemented yet.")
