pyproject.toml
[tool.poetry]
name = "ppt-translator-agent"
version = "0.1.0"
description = "Google ADK agent that translates PowerPoint presentations via Azure OpenAI"
authors = []
readme = "README.md"
packages = [{ include = "ppt_translator" }]

[tool.poetry.dependencies]
python = "^3.11"
google-adk = ">=1.0.0"
litellm = ">=1.40.0"
openai = ">=1.30.0"
python-pptx = ">=1.0.0"
python-dotenv = ">=1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
